reuse get_before_request_config also in before validate requests and clear g after each request in the tests to ensure fresh context
